Move RenderViewport to its own file
It doesn't need to be in the same file as RenderSliver.
